DEV-41: rename files
