+++ src/pages/PostPage/PostPage.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Comments from '../../components/Comments';
-import { DispatchProps, StateProps } from './BlogPostContainer';
+import { DispatchProps, StateProps } from './PostPageContainer';
 
 export interface OwnProps {}
 
 type Props = StateProps & DispatchProps & OwnProps;
 
-function BlogPost({ post, addComment }: Props) {
+function PostPage({ post, addComment }: Props) {
   if (!post) return <h1>404 page (in dev)</h1>;
   return (
     <Comments
@@ -16,4 +16,4 @@ function BlogPost({ post, addComment }: Props) {
   );
 }
 
-export default BlogPost;
+export default PostPage;
